Move namespace module to core module.
