+++ console/src/test/java/com/alibaba/nacos/console/service/NamespaceOperationServiceTest.java
@@ -21,8 +21,8 @@ import com.alibaba.nacos.api.exception.api.NacosApiException;
 import com.alibaba.nacos.config.server.model.TenantInfo;
 import com.alibaba.nacos.config.server.service.repository.CommonPersistService;
 import com.alibaba.nacos.config.server.service.repository.ConfigInfoPersistService;
-import com.alibaba.nacos.console.enums.NamespaceTypeEnum;
-import com.alibaba.nacos.console.model.Namespace;
+import com.alibaba.nacos.core.namespace.model.NamespaceTypeEnum;
+import com.alibaba.nacos.core.namespace.model.Namespace;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
