+++ core/src/main/java/com/alibaba/nacos/core/namespace/model/Namespace.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.console.model;
+package com.alibaba.nacos.core.namespace.model;
 
 /**
  * Namespace.
@@ -34,7 +34,7 @@ public class Namespace {
     private int configCount;
     
     /**
-     * see {@link com.alibaba.nacos.console.enums.NamespaceTypeEnum}.
+     * see {@link NamespaceTypeEnum}.
      */
     private int type;
